// src/api/generated/model/clientMeResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */

export interface ClientMeResponse {
	userId?: string;
	name?: string;
	email?: string;
	profilePhoto?: string;
	cpf?: string;
	dateOfBirth?: string;
}